+++ tests/LaravelModulesServiceProviderTest.php
@@ -2,13 +2,14 @@
 
 namespace Nwidart\Modules\Tests;
 
-use Nwidart\Modules\Laravel\Repository;
+use Nwidart\Modules\Contracts\RepositoryInterface;
 
 class LaravelModulesServiceProviderTest extends BaseTestCase
 {
     /** @test */
     public function it_binds_modules_key_to_repository_class()
     {
-        $this->assertInstanceOf(Repository::class, app('modules'));
+        $this->assertInstanceOf(RepositoryInterface::class, app(RepositoryInterface::class));
+        $this->assertInstanceOf(RepositoryInterface::class, app('modules'));
     }
 }